requirements.txt
discord.py==2.3.2
pandas
pytz
gspread
google-auth
google-auth-oauthlib
google-auth-httplib2
oauth2client
requests
